requirements.txt
pandas
requests
yfinance
matplotlib
pytz
python-binance
scipy
websockets
plotly